pyproject.toml
[project]
name = "apeiron"
version = "0.1.0"
description = "A multitask AI agent designed to streamline and automate operational workflows"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "click>=8.1.8",
  "fastapi>=0.122.0",
  "langchain-core>=0.3.40",
  "langchain-google-genai>=2.1.0",
  "langchain-mistralai>=0.2.7",
  "langchain>=0.3.19",
  "langgraph>=0.3.2",
  "langmem>=0.0.17",
  "mistral-common[sentencepiece]>=1.5.3",
  "py-cord[speed,voice]>=2.6.1",
  "pydantic>=2.10.6",
  "pyyaml>=6.0.2",
  "tiktoken>=0.9.0",
  "transformers>=4.49.0",
  "uvicorn>=0.34.0",
]

[dependency-groups]
dev = ["ruff>=0.9.9"]

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
